+++ envelope/__init__.py
@@ -2,4 +2,5 @@ import sys
 from .envelope import Envelope
 
 sys.modules[__name__] = Envelope  # to allow dynamic module call `import envelope; envelope(...)`
-Envelope.envelope = Envelope  # to support static inspection autocompletion
+envelope = Envelope  # to support static inspection autocompletion `from envelope import envelope`
+Envelope.envelope = Envelope  # point `from envelope import envelope` at working `import envelope` at runtime