+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 0.9.2 (unreleased)
 - CLI: if nothing to do, assume parameters are a bone of an e-mail message to produce output
+- check GPG signing failed
+- fix GnuPG home
+- fix encrypted subject 
+- SMTP supports INI file
+- `check` will print out DNS information of the sender's domain
+- CLI: metavar display names added
+- attach_key flag added 
 
 ## 0.9.1
 - signing, encrypting, sending